Started test refactor
