+++ test/index.spec.js
@@ -3,14 +3,13 @@ import { expect } from 'chai'
 
 import sinon from 'sinon'
 import { LawReg } from '../src/index.js'
-import * as log from 'loglevel'
 import Util from '../src/util'
 
 import awb from './flint-example-awb'
 import IdentityUtil from '../src/identity_util'
+import { setupLogging } from './logging'
 
-// Adjusting log level for debugging can be done here, or in specific tests that need more finegrained logging during development
-log.getLogger('disciplLawReg').setLevel('debug')
+setupLogging()
 
 const lawReg = new LawReg()
 
@@ -2064,1373 +2063,6 @@ describe('discipl-law-reg', () => {
         'value': '5'
       }
       )
-    })
-  })
-
-  describe('CREATE expression', async function () {
-    it('should take an action if the fact and operands are created', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<aanvraag kinderbijslag>>',
-            'actor': '[ouder]',
-            'recipient': '[minister]',
-            'object': '[verzoek]',
-            'create': [
-              '[aanvraag]'
-            ]
-          },
-          {
-            'act': '<<bedrag vaststellen>>',
-            'actor': '[ouder]',
-            'recipient': '[ouder]',
-            'object': '[bedrag]',
-            'create': [
-              '[bedrag]'
-            ]
-          },
-          {
-            'act': '<<aanvraag kinderbijslag toekennen>>',
-            'actor': '[minister]',
-            'object': '[aanvraag]',
-            'recipient': '[ouder]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[bedrag]'
-              ]
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[ouder]': true, '[minister]': true, '[verzoek]': true, '[bedrag]': 100 }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['ouder', 'minister'], { 'ouder': '[ouder]', 'minister': '[minister]' })
-      const acts = [
-        {
-          'act': '<<bedrag vaststellen>>',
-          'actor': 'ouder'
-        },
-        {
-          'act': '<<aanvraag kinderbijslag>>',
-          'actor': 'ouder'
-        },
-        {
-          'act': '<<aanvraag kinderbijslag toekennen>>',
-          'actor': 'minister'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(acts).to.deep.include({ 'act': '<<bedrag vaststellen>>', 'actor': 'ouder' })
-      expect(model.acts[0]).to.deep.include({ 'create': ['[aanvraag]'] })
-      expect(model.acts[1]).to.deep.include({ 'create': ['[bedrag]'] })
-      expect(errorMessage).to.equal('')
-    })
-
-    it('should not take an action if the fact is supplied', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<aanvraag kinderbijslag toekennen>>',
-            'actor': '[minister]',
-            'object': '[aanvraag]',
-            'recipient': '[ouder]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE'
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[ouder]': true, '[minister]': true }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['minister'], { 'minister': '[minister]' })
-      const acts = [
-        {
-          'act': '<<aanvraag kinderbijslag toekennen>>',
-          'actor': 'minister'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(model.acts[0]).to.not.deep.include({ 'create': ['[aanvraag]'] })
-      expect(errorMessage).to.equal('Action <<aanvraag kinderbijslag toekennen>> is not allowed due to object')
-    })
-
-    it('should not take an action if a fact given as operand is not given', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<bedrag vaststellen>>',
-            'actor': '[ouder]',
-            'recipient': '[ouder]',
-            'object': '[bedrag]',
-            'create': [
-              '[bedrag]'
-            ]
-          },
-          {
-            'act': '<<aanvraag kinderbijslag>>',
-            'actor': '[ouder]',
-            'recipient': '[minister]',
-            'object': '[verzoek]',
-            'create': [
-              '[aanvraag]'
-            ]
-          },
-          {
-            'act': '<<aanvraag kinderbijslag toekennen>>',
-            'actor': '[minister]',
-            'object': '[aanvraag]',
-            'recipient': '[ouder]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[bedrag]'
-              ]
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[ouder]': true, '[minister]': true, '[verzoek]': true }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['ouder', 'minister'], { 'ouder': '[ouder]', 'minister': '[minister]' })
-      const acts = [
-        {
-          'act': '<<aanvraag kinderbijslag>>',
-          'actor': 'ouder'
-        },
-        {
-          'act': '<<aanvraag kinderbijslag toekennen>>',
-          'actor': 'minister'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(Object.keys(completeFacts)).to.not.include('[bedrag]')
-      expect(acts).to.not.deep.include({ 'act': '<<bedrag vaststellen>>', 'actor': 'ouder' })
-      expect(errorMessage).to.equal('Action <<aanvraag kinderbijslag toekennen>> is not allowed due to object')
-    })
-
-    it('should allow OR expressions', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<aanvragen kinderbijslag>>',
-            'actor': '[ouder]',
-            'recipient': '[ambtenaar]',
-            'object': '[verzoek]',
-            'preconditions': '[maximaal]',
-            'create': [
-              '[aanvraag]'
-            ]
-          },
-          {
-            'act': '<<aanvraag kinderbijslag toekennen>>',
-            'actor': '[ambtenaar]',
-            'object': '[aanvraag]',
-            'recipient': '[ouder]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                {
-                  'expression': 'OR',
-                  'operands': [
-                    '[bedrag]',
-                    '[maximaal]'
-                  ]
-                }
-              ]
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[ouder]': true, '[ambtenaar]': true, '[verzoek]': true, '[maximaal]': true }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['ambtenaar', 'ouder'], completeFacts)
-      const acts = [
-        {
-          'act': '<<aanvragen kinderbijslag>>',
-          'actor': 'ouder'
-        },
-        {
-          'act': '<<aanvraag kinderbijslag toekennen>>',
-          'actor': 'ambtenaar'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(errorMessage).to.equal('')
-    })
-  })
-
-  describe('PROJECTION expression', async function () {
-    it('should be able to have multiple of the same type of actor', async () => {
-      const subsidieModel = {
-        'acts': [
-          {
-            'act': '<<subsidie aanvragen>>',
-            'actor': '[burger]',
-            'action': '[aanvragen]',
-            'object': '[verzoek]',
-            'recipient': '[ambtenaar]',
-            'preconditions': {
-              'expression': 'AND',
-              'operands': [
-                '[bsn]',
-                '[bedrag]'
-              ]
-            },
-            'create': [
-              '[aanvraag]'
-            ],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          },
-          {
-            'act': '<<subsidie aanvraag intrekken>>',
-            'actor': '[burger with aanvraag]',
-            'action': '[intrekken]',
-            'object': '[aanvraag]',
-            'recipient': '[ambtenaar]',
-            'preconditions': '[]',
-            'create': [],
-            'terminate': [
-              '[aanvraag]'
-            ],
-            'sources': [],
-            'explanation': ''
-          },
-          {
-            'act': '<<subsidie aanvraag toekennen>>',
-            'actor': '[ambtenaar]',
-            'action': '[toekennen]',
-            'object': '[aanvraag]',
-            'recipient': '[burger]',
-            'preconditions': {
-              'expression': 'LESS_THAN',
-              'operands': [
-                '[bedrag projection]',
-                {
-                  'expression': 'LITERAL',
-                  'operand': 500
-                }
-              ]
-            },
-            'create': [],
-            'terminate': [
-              '[aanvraag]'
-            ],
-            'sources': [],
-            'explanation': ''
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[bedrag]',
-            'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[aanvraag]',
-            'explanation': '',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[bedrag]',
-                '[burger]'
-              ]
-            },
-            'sources': []
-          },
-          {
-            'fact': '[bedrag projection]',
-            'explanation': '',
-            'function': {
-              'expression': 'PROJECTION',
-              'context': [
-                '[aanvraag]'
-              ],
-              'fact': '[bedrag]'
-            },
-            'sources': []
-          },
-          {
-            'fact': '[burger with aanvraag]',
-            'explanation': '',
-            'function': {
-              'expression': 'PROJECTION',
-              'context': [
-                '[aanvraag]'
-              ],
-              'fact': '[burger]'
-            },
-            'sources': []
-          },
-          {
-            'fact': '[burger]',
-            'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[verzoek]',
-            'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[ambtenaar]',
-            'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[aanvragen]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[toekennen]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          }
-        ],
-        'duties': []
-      }
-      const core = lawReg.getAbundanceService().getCoreAPI()
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(
-        subsidieModel,
-        ['burger1', 'burger2', 'ambtenaar'],
-        { '[ambtenaar]': 'ambtenaar', '[burger]': ['burger1', 'burger2'] }
-      )
-
-      const needLink = await core.claim(ssids['burger1'], {
-        'need': {
-          'act': '<<subsidie aanvragen>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const factResolver = (fact) => {
-        if (fact === '[bedrag]') { return 50 }
-        if (fact === '[bsn]') { return '605211619' }
-        return fact === '[verzoek]'
-      }
-
-      const availableActs1 = (await lawReg.getAvailableActsWithResolver(needLink, ssids['burger1'], factResolver)).map((actInfo) => actInfo.act)
-      expect(availableActs1).to.deep.equal(['<<subsidie aanvragen>>'])
-
-      const actionLink = await lawReg.take(ssids['burger1'], needLink, '<<subsidie aanvragen>>', factResolver)
-      const availableActs2 = (await lawReg.getAvailableActs(actionLink, ssids['burger1'])).map((actInfo) => actInfo.act)
-      expect(availableActs2).to.deep.equal(['<<subsidie aanvraag intrekken>>'])
-
-      const burger2Acts = (await lawReg.getAvailableActs(actionLink, ssids['burger2'])).map((actInfo) => actInfo.act)
-      expect(burger2Acts).to.deep.equal([])
-
-      const ambtenaarActs = (await lawReg.getAvailableActs(actionLink, ssids['ambtenaar'])).map((actInfo) => actInfo.act)
-      expect(ambtenaarActs).to.deep.equal(['<<subsidie aanvraag toekennen>>'])
-    })
-
-    const subsidieModel = {
-      'acts': [
-        {
-          'act': '<<subsidie aanvragen>>',
-          'actor': '[burger]',
-          'action': '[aanvragen]',
-          'object': '[verzoek]',
-          'recipient': '[ambtenaar]',
-          'preconditions': '[bedrag]',
-          'create': [
-            '[aanvraag]'
-          ],
-          'terminate': [],
-          'sources': [],
-          'explanation': ''
-        },
-        {
-          'act': '<<subsidie aanvraag toekennen>>',
-          'actor': '[ambtenaar]',
-          'action': '[toekennen]',
-          'object': '[aanvraag]',
-          'recipient': '[burger]',
-          'preconditions': {
-            'expression': 'LESS_THAN',
-            'operands': [
-              '[bedrag projection]',
-              {
-                'expression': 'LITERAL',
-                'operand': 500
-              }
-            ]
-          },
-          'create': [],
-          'terminate': [
-            '[aanvraag]'
-          ],
-          'sources': [],
-          'explanation': ''
-        }
-      ],
-      'facts': [
-        {
-          'fact': '[bedrag]',
-          'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-          'function': '[]',
-          'sources': []
-        },
-        {
-          'fact': '[aanvraag]',
-          'explanation': '',
-          'function': {
-            'expression': 'CREATE',
-            'operands': [
-              '[bedrag]'
-            ]
-          },
-          'sources': []
-        },
-        {
-          'fact': '[bedrag projection]',
-          'explanation': '',
-          'function': {
-            'expression': 'PROJECTION',
-            'context': [
-              '[aanvraag]'
-            ],
-            'fact': '[bedrag]'
-          },
-          'sources': []
-        },
-        {
-          'fact': '[burger]',
-          'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-          'function': '[]',
-          'sources': []
-        },
-        {
-          'fact': '[verzoek]',
-          'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-          'function': '[]',
-          'sources': []
-        },
-        {
-          'fact': '[ambtenaar]',
-          'explanation': 'GENERATED: This fact was generated during the \'Import From Json Action\'',
-          'function': '[]',
-          'sources': []
-        },
-        {
-          'fact': '[aanvragen]',
-          'explanation': '',
-          'function': '[]',
-          'sources': []
-        },
-        {
-          'fact': '[toekennen]',
-          'explanation': '',
-          'function': '[]',
-          'sources': []
-        }
-      ],
-      'duties': []
-    }
-
-    it('should call getPotentialActs multiple times without breaking PROJECTION expressions', async () => {
-      const core = lawReg.getAbundanceService().getCoreAPI()
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(subsidieModel, ['burger', 'ambtenaar'], { '[ambtenaar]': 'ambtenaar', '[burger]': 'burger' })
-
-      const needLink = await core.claim(ssids['burger'], {
-        'need': {
-          'act': '<<subsidie aanvragen>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const potentialActs = (await lawReg.getPotentialActs(needLink, ssids['burger'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(potentialActs).to.deep.equal(['<<subsidie aanvragen>>'])
-
-      const potentialActs2 = (await lawReg.getPotentialActs(needLink, ssids['ambtenaar'], [], [])).map((actInfo) => actInfo.act)
-      expect(potentialActs2).to.deep.equal([])
-    })
-
-    it('should get the projected property', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<subsidie aanvragen>>',
-            'actor': '[burger]',
-            'recipient': '[ambtenaar]',
-            'object': '[verzoek]',
-            'preconditions': '[bedrag]',
-            'create': [
-              '[aanvraag]'
-            ]
-          },
-          {
-            'act': '<<subsidie aanvraag toekennen>>',
-            'actor': '[ambtenaar]',
-            'object': '[aanvraag]',
-            'preconditions': {
-              'expression': 'EQUAL',
-              'operands': [
-                {
-                  'expression': 'PROJECTION',
-                  'context': [
-                    '[aanvraag]'
-                  ],
-                  'fact': '[bedrag]'
-                },
-                {
-                  'expression': 'LITERAL',
-                  'operand': 500
-                }
-              ]
-            },
-            'recipient': '[burger]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[bedrag]'
-              ]
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[burger]': true, '[ambtenaar]': true, '[verzoek]': true, '[bedrag]': 500 }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['burger', 'ambtenaar'], { 'burger': '[burger]', 'ambtenaar': '[ambtenaar]' })
-      const acts = [
-        {
-          'act': '<<subsidie aanvragen>>',
-          'actor': 'burger'
-        },
-        {
-          'act': '<<subsidie aanvraag toekennen>>',
-          'actor': 'ambtenaar'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(completeFacts).to.deep.include({ '[bedrag]': 500 })
-      expect(errorMessage).to.equal('')
-    })
-
-    it('should project a series of CREATE facts', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<persoonlijk gegevens invullen>>',
-            'actor': '[burger]',
-            'recipient': '[ambtenaar]',
-            'object': '[verzoek]',
-            'preconditions': '[naam]',
-            'create': [
-              '[persoonlijke gegevens]'
-            ]
-          },
-          {
-            'act': '<<subsidie aanvragen>>',
-            'actor': '[burger]',
-            'recipient': '[ambtenaar]',
-            'object': '[verzoek]',
-            'preconditions': {
-              'expression': 'AND',
-              'operands': [
-                '[persoonlijke gegevens]',
-                '[bedrag]'
-              ]
-            },
-            'create': [
-              '[aanvraag]'
-            ]
-          },
-          {
-            'act': '<<subsidie aanvraag toekennen>>',
-            'actor': '[ambtenaar]',
-            'object': '[aanvraag]',
-            'preconditions': {
-              'expression': 'EQUAL',
-              'operands': [
-                {
-                  'expression': 'PROJECTION',
-                  'context': [
-                    '[aanvraag]',
-                    '[persoonlijke gegevens]'
-                  ],
-                  'fact': '[naam]'
-                },
-                {
-                  'expression': 'LITERAL',
-                  'operand': 'Discipl'
-                }
-              ]
-            },
-            'recipient': '[burger]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[persoonlijke gegevens]',
-                '[bedrag]'
-              ]
-            }
-          },
-          {
-            'fact': '[persoonlijke gegevens]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[naam]'
-              ]
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[burger]': true, '[ambtenaar]': true, '[verzoek]': true, '[naam]': 'Discipl', '[bedrag]': 500 }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['burger', 'ambtenaar'], { 'burger': '[burger]', 'ambtenaar': '[ambtenaar]' })
-      const acts = [
-        {
-          'act': '<<persoonlijk gegevens invullen>>',
-          'actor': 'burger'
-        },
-        {
-          'act': '<<subsidie aanvragen>>',
-          'actor': 'burger'
-        },
-        {
-          'act': '<<subsidie aanvraag toekennen>>',
-          'actor': 'ambtenaar'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(completeFacts).to.deep.include({ '[bedrag]': 500, '[naam]': 'Discipl' })
-      expect(errorMessage).to.equal('')
-    })
-
-    it('should not allow an act if the projection failed', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<subsidie aanvraag toekennen>>',
-            'actor': '[ambtenaar]',
-            'object': '[aanvraag]',
-            'preconditions': {
-              'expression': 'EQUAL',
-              'operands': [
-                {
-                  'expression': 'PROJECTION',
-                  'context': [
-                    '[aanvraag]'
-                  ],
-                  'fact': '[niet bestaand bedrag]'
-                },
-                {
-                  'expression': 'LITERAL',
-                  'operand': 500
-                }
-              ]
-            },
-            'recipient': '[burger]'
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[aanvraag]',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[bedrag]'
-              ]
-            }
-          }
-        ],
-        'duties': []
-      }
-
-      const completeFacts = { '[burger]': true, '[ambtenaar]': true, '[verzoek]': true, '[bedrag]': 500 }
-      const util = new Util(lawReg)
-
-      const { ssids, modelLink } = await util.setupModel(model, ['burger', 'ambtenaar'], { 'burger': '[burger]', 'ambtenaar': '[ambtenaar]' })
-      const acts = [
-        {
-          'act': '<<subsidie aanvraag toekennen>>',
-          'actor': 'ambtenaar'
-        }
-      ]
-
-      let errorMessage = ''
-      try {
-        await util.scenarioTest(ssids, modelLink, acts, completeFacts)
-      } catch (e) {
-        errorMessage = e.message
-      }
-
-      expect(completeFacts).to.deep.include({ '[bedrag]': 500 })
-      expect(errorMessage).to.equal('Action <<subsidie aanvraag toekennen>> is not allowed due to object')
-    })
-
-    it('should be able to take an action after object is created from other action', async () => {
-      const core = lawReg.getAbundanceService().getCoreAPI()
-
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(subsidieModel, ['burger', 'ambtenaar'], { '[ambtenaar]': 'ambtenaar', '[burger]': 'burger' })
-
-      const needSsid = await core.newSsid('ephemeral')
-
-      await core.allow(needSsid)
-
-      const needLink = await core.claim(needSsid, {
-        'need': {
-          'act': '<<subsidie aanvragen>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const factResolver = (fact) => {
-        if (fact === '[bedrag]') {
-          return 50
-        }
-        return fact === '[verzoek]'
-      }
-
-      const actionLink = await lawReg.take(ssids['burger'], needLink, '<<subsidie aanvragen>>', factResolver)
-
-      const ambtenaarAvailableActs = (await lawReg.getAvailableActs(actionLink, ssids['ambtenaar'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(ambtenaarAvailableActs).to.deep.equal(['<<subsidie aanvraag toekennen>>'])
-
-      const burgerPotentialActs = (await lawReg.getPotentialActs(actionLink, ssids['burger'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(burgerPotentialActs).to.deep.equal(['<<subsidie aanvragen>>'])
-
-      const burgerAvailableActs = (await lawReg.getAvailableActs(actionLink, ssids['burger'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(burgerAvailableActs).to.deep.equal([])
-
-      const actionLink2 = await lawReg.take(ssids['ambtenaar'], actionLink, '<<subsidie aanvraag toekennen>>', factResolver)
-
-      const ambtenaarAvailableActsAfterAanvraagToegekend = (await lawReg.getAvailableActs(actionLink2, ssids['ambtenaar'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(ambtenaarAvailableActsAfterAanvraagToegekend).to.deep.equal([])
-    })
-
-    it('EQUAL should return undefined when a PROJECTION expression returns undefined', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<bake pancake>>',
-            'actor': '[actor1]',
-            'action': '[bake]',
-            'object': '[batter]',
-            'recipient': '[actor2]',
-            'preconditions': '[ingredients]',
-            'create': [
-              '[baked pancake]'
-            ],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          },
-          {
-            'act': '<<eat pancake>>',
-            'actor': '[actor1]',
-            'action': '[eat]',
-            'object': '[baked pancake]',
-            'recipient': '[actor2]',
-            'preconditions': {
-              'expression': 'EQUAL',
-              'operands': [
-                {
-                  'expression': 'LITERAL',
-                  'operand': 'apple'
-                },
-                {
-                  'expression': 'PROJECTION',
-                  'context': [
-                    '[baked pancake]'
-                  ],
-                  'fact': '[ingredients]'
-                }
-              ]
-            },
-            'create': [],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[actor1]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[bake]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[batter]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[actor2]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[ingredients]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[baked pancake]',
-            'explanation': '',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[ingredients]'
-              ]
-            },
-            'sources': []
-          },
-          {
-            'fact': '[eat]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          }
-        ],
-        'duties': []
-      }
-
-      const core = lawReg.getAbundanceService().getCoreAPI()
-
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(model, ['Actor1'], { '[actor1]': 'Actor1' })
-
-      const needSsid = await core.newSsid('ephemeral')
-
-      await core.allow(needSsid)
-
-      const needLink = await core.claim(needSsid, {
-        'need': {
-          'act': '<<bake pancake>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const appleFactResolver = (fact) => {
-        if (fact === '[ingredients]') {
-          return 'apple'
-        }
-        return fact === '[batter]' || fact === '[actor2]'
-      }
-
-      const actionLink = await lawReg.take(ssids['Actor1'], needLink, '<<bake pancake>>', appleFactResolver)
-
-      const actor1Acts = (await lawReg.getPotentialActs(actionLink, ssids['Actor1'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts).to.include('<<eat pancake>>')
-
-      const blueberryFactResolver = (fact) => {
-        if (fact === '[ingredients]') {
-          return 'blueberry'
-        }
-        return fact === '[batter]' || fact === '[actor2]'
-      }
-
-      const actionLink1 = await lawReg.take(ssids['Actor1'], actionLink, '<<bake pancake>>', blueberryFactResolver)
-
-      const actor1Acts1 = (await lawReg.getPotentialActs(actionLink1, ssids['Actor1'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts1).to.include('<<eat pancake>>')
-    })
-
-    it('LESS THAN should return undefined when a PROJECTION expression returns undefined', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<give a number>>',
-            'actor': '[actor1]',
-            'action': 'action',
-            'object': '[calculator]',
-            'recipient': '[actor1]',
-            'preconditions': '[number]',
-            'create': [
-              '[paper]'
-            ],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          },
-          {
-            'act': '<<accept number>>',
-            'actor': '[actor1]',
-            'action': 'accept',
-            'object': '[calculator]',
-            'recipient': '[actor1]',
-            'preconditions': {
-              'expression': 'LESS_THAN',
-              'operands': [
-                {
-                  'expression': 'LITERAL',
-                  'operand': 1
-                },
-                {
-                  'expression': 'PROJECTION',
-                  'context': [
-                    '[paper]'
-                  ],
-                  'fact': '[number]'
-                }
-              ]
-            },
-            'create': [],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[actor1]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[calculator]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[paper]',
-            'explanation': '',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[number]'
-              ]
-            },
-            'sources': []
-          },
-          {
-            'fact': '[number]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          }
-        ],
-        'duties': []
-      }
-
-      const core = lawReg.getAbundanceService().getCoreAPI()
-
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(model, ['Actor'], { '[actor1]': 'Actor' })
-
-      const needSsid = await core.newSsid('ephemeral')
-
-      await core.allow(needSsid)
-
-      const needLink = await core.claim(needSsid, {
-        'need': {
-          'act': '<<give a number>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const factResolver = (fact) => {
-        if (fact === '[number]') {
-          return 10
-        }
-        return fact === '[calculator]'
-      }
-
-      const actionLink = await lawReg.take(ssids['Actor'], needLink, '<<give a number>>', factResolver)
-
-      const actor1Acts = (await lawReg.getPotentialActs(actionLink, ssids['Actor'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts).to.include('<<accept number>>')
-
-      const factResolver2 = (fact) => {
-        if (fact === '[number]') {
-          return 0
-        }
-        return fact === '[calculator]'
-      }
-
-      const actionLink1 = await lawReg.take(ssids['Actor'], needLink, '<<give a number>>', factResolver2)
-
-      const actor1Acts1 = (await lawReg.getPotentialActs(actionLink1, ssids['Actor'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts1).to.include('<<accept number>>')
-    })
-
-    it('SUM should return undefined when a PROJECTION expression returns undefined', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<give a number>>',
-            'actor': '[actor1]',
-            'action': 'action',
-            'object': '[calculator]',
-            'recipient': '[actor1]',
-            'preconditions': '[number]',
-            'create': [
-              '[paper]'
-            ],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          },
-          {
-            'act': '<<accept number>>',
-            'actor': '[actor1]',
-            'action': 'accept',
-            'object': '[calculator]',
-            'recipient': '[actor1]',
-            'preconditions': {
-              'expression': 'EQUAL',
-              'operands': [
-                {
-                  'expression': 'LITERAL',
-                  'operand': 11
-                },
-                {
-                  'expression': 'SUM',
-                  'operands': [
-                    {
-                      'expression': 'LITERAL',
-                      'operand': 1
-                    },
-                    {
-                      'expression': 'PROJECTION',
-                      'context': [
-                        '[paper]'
-                      ],
-                      'fact': '[number]'
-                    }
-                  ]
-                }
-
-              ]
-            },
-            'create': [],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[actor1]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[calculator]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[paper]',
-            'explanation': '',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[number]'
-              ]
-            },
-            'sources': []
-          },
-          {
-            'fact': '[number]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          }
-        ],
-        'duties': []
-      }
-
-      const core = lawReg.getAbundanceService().getCoreAPI()
-
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(model, ['Actor'], { '[actor1]': 'Actor' })
-
-      const needSsid = await core.newSsid('ephemeral')
-
-      await core.allow(needSsid)
-
-      const needLink = await core.claim(needSsid, {
-        'need': {
-          'act': '<<give a number>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const factResolver = (fact) => {
-        if (fact === '[number]') {
-          return 10
-        }
-        return fact === '[calculator]'
-      }
-
-      const actionLink = await lawReg.take(ssids['Actor'], needLink, '<<give a number>>', factResolver)
-
-      const actor1Acts = (await lawReg.getPotentialActs(actionLink, ssids['Actor'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts).to.include('<<accept number>>')
-
-      const factResolver2 = (fact) => {
-        if (fact === '[number]') {
-          return 0
-        }
-        return fact === '[calculator]'
-      }
-
-      const actionLink1 = await lawReg.take(ssids['Actor'], needLink, '<<give a number>>', factResolver2)
-
-      const actor1Acts1 = (await lawReg.getPotentialActs(actionLink1, ssids['Actor'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts1).to.include('<<accept number>>')
-    })
-
-    it('PRODUCT should return undefined when a PROJECTION expression returns undefined', async () => {
-      const model = {
-        'acts': [
-          {
-            'act': '<<give a number>>',
-            'actor': '[actor1]',
-            'action': 'action',
-            'object': '[calculator]',
-            'recipient': '[actor1]',
-            'preconditions': '[number]',
-            'create': [
-              '[paper]'
-            ],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          },
-          {
-            'act': '<<accept number>>',
-            'actor': '[actor1]',
-            'action': 'accept',
-            'object': '[calculator]',
-            'recipient': '[actor1]',
-            'preconditions': {
-              'expression': 'EQUAL',
-              'operands': [
-                {
-                  'expression': 'LITERAL',
-                  'operand': 20
-                },
-                {
-                  'expression': 'PRODUCT',
-                  'operands': [
-                    {
-                      'expression': 'LITERAL',
-                      'operand': 2
-                    },
-                    {
-                      'expression': 'PROJECTION',
-                      'context': [
-                        '[paper]'
-                      ],
-                      'fact': '[number]'
-                    }
-                  ]
-                }
-
-              ]
-            },
-            'create': [],
-            'terminate': [],
-            'sources': [],
-            'explanation': ''
-          }
-        ],
-        'facts': [
-          {
-            'fact': '[actor1]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[calculator]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          },
-          {
-            'fact': '[paper]',
-            'explanation': '',
-            'function': {
-              'expression': 'CREATE',
-              'operands': [
-                '[number]'
-              ]
-            },
-            'sources': []
-          },
-          {
-            'fact': '[number]',
-            'explanation': '',
-            'function': '[]',
-            'sources': []
-          }
-        ],
-        'duties': []
-      }
-
-      const core = lawReg.getAbundanceService().getCoreAPI()
-
-      const util = new Util(lawReg)
-      const { ssids, modelLink } = await util.setupModel(model, ['Actor'], { '[actor1]': 'Actor' })
-
-      const needSsid = await core.newSsid('ephemeral')
-
-      await core.allow(needSsid)
-
-      const needLink = await core.claim(needSsid, {
-        'need': {
-          'act': '<<give a number>>',
-          'DISCIPL_FLINT_MODEL_LINK': modelLink
-        }
-      })
-
-      const factResolver = (fact) => {
-        if (fact === '[number]') {
-          return 10
-        }
-        return fact === '[calculator]'
-      }
-
-      const actionLink = await lawReg.take(ssids['Actor'], needLink, '<<give a number>>', factResolver)
-
-      const actor1Acts = (await lawReg.getPotentialActs(actionLink, ssids['Actor'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts).to.include('<<accept number>>')
-
-      const factResolver2 = (fact) => {
-        if (fact === '[number]') {
-          return 0
-        }
-        return fact === '[calculator]'
-      }
-
-      const actionLink1 = await lawReg.take(ssids['Actor'], needLink, '<<give a number>>', factResolver2)
-
-      const actor1Acts1 = (await lawReg.getPotentialActs(actionLink1, ssids['Actor'], [], [])).map((actInfo) => actInfo.act)
-
-      expect(actor1Acts1).to.include('<<accept number>>')
     })
   })
 })
